Implement memory access bound check with hardware trap for 64-bit platforms (#293)
Also implement native stack overflow check with hardware trap for 64-bit platforms
Refine classic interpreter and fast interpreter to improve performance
Update document

+++ core/shared/platform/include/platform_api_vmcore.h
@@ -110,9 +110,9 @@ enum {
     MMAP_MAP_FIXED = 2
 };
 
-void *os_mmap(void *hint, unsigned int size, int prot, int flags);
-void os_munmap(void *addr, uint32 size);
-int os_mprotect(void *addr, uint32 size, int prot);
+void *os_mmap(void *hint, size_t size, int prot, int flags);
+void os_munmap(void *addr, size_t size);
+int os_mprotect(void *addr, size_t size, int prot);
 
 /**
  * Flush cpu data cache, in some CPUs, after applying relocation to the
